refactor: extract auth feature, composables, and help validation schemas (#511)
