+++ apps/container/src/components/layout/AppBar/AppBar.vue
@@ -165,8 +165,8 @@ import moonIcon from '@/assets/icons/moon.svg';
 import sunIcon from '@/assets/icons/sun.svg';
 import logoDark from '@/assets/logo.svg';
 import logoLight from '@/assets/logo_white.svg';
-import { UserMenu } from '@/components/UserMenu';
 import { applyChartsTheme } from '@/config/highcharts';
+import { UserMenu } from '@/features/auth/components/UserMenu';
 import { eventTracker } from '@/helpers/EventTracker';
 import { WebEvent } from '@/helpers/WebEvent';
 import { useAuthStore } from '@/stores/auth';
