+++ apps/container/src/composables/cleanUsername.ts
